requirements.txt
fastapi==0.95.0
python-multipart==0.0.6
python-jose==3.3.0
sqlalchemy==1.4.23
fernet==1.0.1